Strategy Scanner from Command
Added a code that get the content of the attached file on discord command.
So you can upload ur strategy and put $analyse in the message.
Python code will save the file in the good rep, then the API will shortly restart, and then discord bot will make the request, give you the result and delete the temp file for the strategy.

This short code is working, but wouldnt work in multiple usage.